requirements.txt
requests_oauthlib
django
django-cas-client
